Changes use of double to BigDecimal to try to make it work with Postgres

+++ conf/evolutions/default/1.sql
@@ -14,7 +14,7 @@ create table raw_utterance (
   id                        bigint not null,
   text                      TEXT,
   timestamp                 TEXT,
-  confidence                double,
+  confidence                NUMERIC,
   constraint pk_raw_utterance primary key (id))
 ;
